Fortifications implemented.
There is only one blue test prefab. The fortification is placed left of the player. Right now, there is a bug that makes it hard to hit the button. You have to click at the top of the button or it presses the mine button for some reason. I will open a new bug for this.